you can get a reminder
In the message.json you can configure a reminder in days.
You have to take care that you message.json is according
to the message.example.json.

Changes where
added structure for birthdayMal and reminderMail to configure
them independent. Added attribute for a reminder, that is send x days before birthday.